chore: move codemod and mutualize test setup
