--- packages/cmf-cqrs/env__tests.js
@@ -1,12 +1,0 @@
-import 'babel-polyfill';
-
-const fetch = jest.fn(
-	(url, config) =>
-		new Promise(resolve => {
-			if (config.response) {
-				return resolve(config.response);
-			}
-			return resolve();
-		}),
-);
-global.fetch = fetch;
