--- packages/containers/env__tests.js
@@ -1,1 +1,0 @@
-import 'babel-polyfill';
