+++ packages/datagrid/package.json
@@ -128,9 +128,6 @@
     "roots": [
       "src"
     ],
-    "setupFiles": [
-      "./env__tests.js"
-    ],
     "transform": {
       ".*": "<rootDir>/../../node_modules/babel-jest"
     },
@@ -138,7 +135,7 @@
     "moduleNameMapper": {
       "^.+\\.(css|scss)$": "<rootDir>/test/styleMock.js"
     },
-    "setupTestFrameworkScriptFile": "<rootDir>/test-setup.js",
+    "setupTestFrameworkScriptFile": "<rootDir>/../../test-setup.js",
     "coveragePathIgnorePatterns": [
       "index.js"
     ]
