--- packages/forms/test-setup.js
@@ -1,4 +1,0 @@
-import { configure } from 'enzyme';
-import Adapter from 'enzyme-adapter-react-15';
-
-configure({ adapter: new Adapter() });
