+++ packages/sagas/package.json
@@ -107,9 +107,7 @@
       "src"
     ],
     "testRegex": "&*\\.test\\.js$",
-    "setupFiles": [
-      "./env__tests.js"
-    ]
+    "setupTestFrameworkScriptFile": "<rootDir>/../../test-setup.js"
   },
   "publishConfig": {
     "access": "public"
