+++ packages/scripts/preset/config/test/test-setup.js
@@ -1,5 +1,16 @@
-require('babel-polyfill');
-const configure = require('enzyme').configure;
-const Adapter = require('enzyme-adapter-react-15');
+import 'babel-polyfill';
+import { configure } from 'enzyme';
+import Adapter from 'enzyme-adapter-react-15';
 
 configure({ adapter: new Adapter() });
+
+const fetch = jest.fn(
+	(url, config) =>
+		new Promise(resolve => {
+			if (config.response) {
+				return resolve(config.response);
+			}
+			return resolve();
+		}),
+);
+global.fetch = fetch;
