+++ scripts/cmf-componentState-deprecation/README.md
