+++ scripts/cmf-componentState-deprecation/componentState_to_cmfConnect.js
